==> ds4j-notebooks/car-crashes-weight-regression/notebooks/04 - Combine VINs and weights.ipynb <==
# Combining car VIN data with vehicle weights

A simple bit of data wrangling.

<p class="reading-options">
  <a class="btn" href="/car-crashes-weight-regression/04-combine-vins-and-weights">
    <i class="fa fa-sm fa-book"></i>
    Read online
  </a>
  <a class="btn" href="/car-crashes-weight-regression/notebooks/04 - Combine VINs and weights.ipynb">
    <i class="fa fa-sm fa-download"></i>
    Download notebook
  </a>
  <a class="btn" href="https://colab.research.google.com/github/littlecolumns/ds4j-notebooks/blob/master/car-crashes-weight-regression/notebooks/04 - Combine VINs and weights.ipynb" target="_new">
    <i class="fa fa-sm fa-laptop"></i>
    Interactive version
  </a>
</p>

### VIN data

vin_df = pd.read_csv("data/vin_data.csv")
vin_df.head()

### Weight data

weights_df = pd.read_csv("data/weights.csv")
weights_df.head()

vins_weights = vin_df.merge(weights_df, 
           left_on=['Make', 'Model', 'ModelYear'], 
           right_on=['Make', 'Model', 'ModelYear'], 
           how='left')
vins_weights.head(5)

vins_weights = vins_weights.dropna(subset=['weight'])
vins_weights.shape

vins_weights.to_csv("data/vins_and_weights.csv", index=False)
